pyproject.toml
[tool.poetry]
name = "bagbag"
version = "0.53.0"
description = "An all in one python library"

license = "MIT"

authors = [
    "Darren"
]

readme = 'README.md'

repository = "https://github.com/darren2046/bagbag"

keywords = ['base', 'library']

exclude = [
  ".git",
  "ident",
  "publish.py",
  "__pycache__",
  ".github",
  "*.ident",
]

[tool.poetry.dependencies]
python = ">=3.9"
loguru = ">=0.6.0"
redis = ">=4.3.4"
# orator = ">=0.9.9"
PyMySQL = ">=1.0.2"
telethon = ">=1.24.0"
langid = ">=1.1.6"
selenium = ">=4.3.0"
Flask = ">=2.1.3"
tqdm = ">=4.64.0"
prometheus-client = ">=0.14.1"
ipdb = ">=0.13.9"
requests = ">=2.28.1"
requests-toolbelt = ">=0.9.1"
pyTelegramBotAPI = ">=4.6.1"
Pillow = ">=9.2.0"
OpenCC = ">=0.2"
windows-curses = { version = ">=2.3.0", platform = "win32" }
random-user-agent = ">=1.0.1"
pythonping = ">=1.1.3"
jieba = ">=0.42.1"
python-dateutil = ">=2.8.2"
schedule = ">=1.1.0"
openpyxl = ">=3.0.10"
lxml = ">=4.9.1"
pygtrans = ">=1.4.0"
shortuuid = ">=1.0.9"
paramiko = ">=2.11.0"
# pycrypto = ">=2.6.1"
pycryptodome = ">=3.15.0"
Markdown = ">=3.4.1"
PyGithub = ">=1.57"
hachoir = ">=3.2.0"
kafka-python = ">=2.0.2"
pybase64 = ">=1.2.3"
listparser = ">=0.19"
feedparser = ">=6.0.10"
markdownify = ">=0.11.6"
pypinyin = ">=0.47.1"
nslookup = ">=1.6.1"
selenium-wire = ">=5.1.0"
tweepy = ">=4.12.1"

# orator 
# Fix bug in mysql8.0: https://github.com/sdispater/orator/issues/361
backpack = ">=0.1"
blinker = ">=1.4"
cleo = ">=0.6"
inflection = ">=0.3"
Faker = ">=0.8"
lazy-object-proxy = ">=1.2"
pendulum = ">=1.4"
pyaml = ">=16.12"
pyyaml = ">=5.1"
Pygments = ">=2.2"
simplejson = ">=3.10"
six = ">=1.10"
wrapt = ">=1.10"

[build-system]
requires = ["poetry>=0.12a3"]
build-backend = "poetry.masonry.api"